pyproject.toml
[build-system]
requires = ["setuptools>=45", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "rag-agent-project"
version = "0.1.0"
description = "A RAG agent with document parsing, vector search, and multi-turn conversation."
dependencies = [
    "python-dotenv",
    "PyPDF2",
    "docx",
    "markdown",
    "faiss-cpu",
    "langchain",
    "langgraph",
    "tiktoken",
    "openai",
    "fastapi",
    "uvicorn"
]

[build-options]
python-version = "3.10"